get_question_items from cache instead of requiring a management command
